refactor: more generic useInputSelect
Remove useInputTable,useSelection,useInputOptionsProvider as generalised useInputSelect and useDataProvider achieve the same.
