+++ lib/composables/__tests__/index.spec.ts
@@ -1,13 +1,12 @@
-import { expect, test } from "vitest"
+import { expect, test } from 'vitest'
 
 test('export all composables except useInput', async () => {
   const index = await import('../index')
+  expect(index).toHaveProperty('useDataProvider')
   expect(index).toHaveProperty('useInputAttrs')
   expect(index).toHaveProperty('useInputBoolean')
-  expect(index).toHaveProperty('useInputText')
   expect(index).toHaveProperty('useInputNumber')
+  expect(index).toHaveProperty('useInputSelect')
+  expect(index).toHaveProperty('useInputText')
   expect(index).toHaveProperty('useInputValidation')
-  expect(index).toHaveProperty('useDataProvider')
-  expect(index).toHaveProperty('useInputTable')
-  expect(index).toHaveProperty('useSelection')
 })
